new additions

+++ src/main/java/com/sbs/hrRecommendation/controllers/recommendationController.java
@@ -1,10 +1,12 @@
 package com.sbs.hrRecommendation.controllers;
-
 import com.sbs.hrRecommendation.models.recommendation;
 import com.sbs.hrRecommendation.repositories.recommendationRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.util.Assert;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,11 +18,13 @@ public class recommendationController {
     @Autowired
     private recommendationRepository recRepository;
 
+    //used to get all the recommendations present in db
     @GetMapping
     public List<recommendation> list() {
         return recRepository.findAll();
     }
 
+    // used to get list of recommendations of a particular user
     @GetMapping
     @RequestMapping("{id}")
     public List<recommendation> listOfRecommendations(@PathVariable Long id) {
@@ -36,24 +40,42 @@ public class recommendationController {
         return result;
     }
 
-    //push data into DB
+    //push recommendation into DB
     @PostMapping
-    public recommendation create(@RequestBody final recommendation Recommendation) {
-        Recommendation.setMyStatus(recommendation.status.SENT);
+    @RequestMapping("/save")
+    public recommendation createDraft(@RequestBody final recommendation Recommendation) {
+        Recommendation.setMyStatus(recommendation.status.DRAFT);
+        return recRepository.saveAndFlush(Recommendation);
+
+    }
+    @PostMapping
+    @RequestMapping("/publish")
+    public recommendation createPublish(@RequestBody final recommendation Recommendation) {
+        Recommendation.setMyStatus(recommendation.status.PENDING);
         return recRepository.saveAndFlush(Recommendation);
 
     }
 
+    //delete a particular recommendation
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE)
     public void delete(@PathVariable Long id) {
-        recRepository.deleteById(id);
+        recommendation rec;
+        String result;
+        rec = recRepository.getOne(id);
+        if (rec.getMyStatus().equals(recommendation.status.DRAFT)){
+            recRepository.deleteById(id);
+        }
+        else
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Status should be only draft");
+
+
     }
 
     //Used to update the record.
-    @RequestMapping(value = "{id}", method = RequestMethod.PUT)
-    public recommendation update(@PathVariable Long id, @RequestBody recommendation Recommendation) {
-        recommendation existingUserProfile = recRepository.getOne(id);
-        BeanUtils.copyProperties(Recommendation, existingUserProfile, "userId");
-        return recRepository.saveAndFlush(existingUserProfile);
-    }
+//    @RequestMapping(value = "{id}", method = RequestMethod.PUT)
+//    public recommendation update(@PathVariable Long id, @RequestBody recommendation Recommendation) {
+//        recommendation existingUserProfile = recRepository.getOne(id);
+//        BeanUtils.copyProperties(Recommendation, existingUserProfile, "userId");
+//        return recRepository.saveAndFlush(existingUserProfile);
+//    }
 }
